changed realex name to realex_offsite
