+++ test/unit/integrations/realex_offsite/realex_offsite_helper_test.rb
@@ -1,6 +1,6 @@
 require 'test_helper'
 
-class RealexHelperTest < Test::Unit::TestCase
+class RealexOffsiteHelperTest < Test::Unit::TestCase
   include OffsitePayments::Integrations
 
   def credentials
@@ -16,7 +16,7 @@ class RealexHelperTest < Test::Unit::TestCase
   end
 
   def setup
-    @helper = Realex::Helper.new('order-500', 'account', order_attributes)
+    @helper = RealexOffsite::Helper.new('order-500', 'account', order_attributes)
   end
 
   def teardown
